scripts/lambda/requirements.txt
requestsrequests
pandas
openpyxl
pyarrow